feat: Implement process-separation IPC architecture for native storj_uplink library
The native storj_uplink (CGO/Go) library now runs in a separate worker
process to avoid runtime conflicts with .NET. All P/Invoke calls are
transparently proxied via a 4-byte length-prefixed JSON IPC protocol
over the worker's stdin/stdout.

Key changes:
- New uplink.NET.Worker console app with WorkerHost, HandleRegistry,
  and OperationDispatcher handling all 27 native operations
- NativeWorkerClient singleton in uplink.NET.Ipc manages the worker
  process lifecycle and serializes requests with SemaphoreSlim(1,1)
- Access.cs, UploadOperation, DownloadOperation, DownloadStream,
  ChunkedUploadOperation refactored to use IPC with long handle IDs
- BucketService, ObjectService, MultipartUploadService refactored to
  use async IPC calls (Task.Run wrappers removed)
- UplinkInterop.cs in main lib stripped to utility methods only
- Worker project output copied to main lib output via MSBuild target
- All public API signatures and test files remain unchanged
- 2 LibraryMetadataTests pass, 57 integration tests skip (no creds)

+++ src/uplink.NET/Models/PartUpload.cs
@@ -1,9 +1,7 @@
-using uplink.NET.Native;
-
 namespace uplink.NET.Models;
 
-/// <summary>Represents a native part-upload handle used in multipart uploads.</summary>
+/// <summary>Represents a part-upload handle used in multipart uploads.</summary>
 public class PartUpload
 {
-    internal nint Handle { get; set; }
+    internal long Handle { get; set; }
 }
